:art: API Wifi calls improved
